Split calico / k8s NP
