+++ lib/backend/model/policy.go
@@ -32,6 +32,10 @@ var (
 	typePolicy  = reflect.TypeOf(Policy{})
 )
 
+const (
+	KindKubernetesNetworkPolicy = "KubernetesNetworkPolicy"
+)
+
 type PolicyKey struct {
 	Name string `json:"-" validate:"required,name"`
 }
